backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
httpx
apscheduler
google-generativeai
python-dotenv
prisma
razorpay
PyJWT
cryptography
pyotp
redis
web3
prometheus-client
prometheus-fastapi-instrumentator
pandas
scikit-learn
xgboost
joblib
networkx
pytest
